move into component ownprops
