+++ webapp/channels/src/components/markdown/markdown.tsx
@@ -3,14 +3,85 @@
 
 import React from 'react';
 
+import type {PostImage, PostType} from '@mattermost/types/posts';
+
 import PostEditedIndicator from 'components/post_view/post_edited_indicator';
 
+import type EmojiMap from 'utils/emoji_map';
 import messageHtmlToComponent from 'utils/message_html_to_component';
+import type {ChannelNamesMap, MentionKey, TextFormattingOptions} from 'utils/text_formatting';
 import {formatText} from 'utils/text_formatting';
 
-import type {PropsFromRedux, OwnProps} from './index';
+import type {PropsFromRedux} from './index';
 
 export type Props = PropsFromRedux & OwnProps;
+
+export type OwnProps = {
+
+    /**
+     * Any additional text formatting options to be used
+     */
+    options?: Partial<TextFormattingOptions>;
+
+    /**
+     * Whether or not to proxy image URLs
+     */
+    proxyImages?: boolean;
+
+    /**
+     * prop for passed down to image component for dimensions
+     */
+    imagesMetadata?: Record<string, PostImage>;
+
+    /**
+     * Post id prop passed down to markdown image
+     */
+    postId?: string;
+
+    /**
+     * When the post is edited this is the timestamp it happened at
+     */
+    editedAt?: number;
+
+    /*
+     * The text to be rendered
+     */
+    message?: string;
+    channelNamesMap?: ChannelNamesMap;
+
+    /*
+     * An array of words that can be used to mention a user
+     */
+    mentionKeys?: MentionKey[];
+
+    /**
+     * Any extra props that should be passed into the image component
+     */
+    imageProps?: object;
+
+    /**
+     * Whether or not to place the LinkTooltip component inside links
+     */
+    hasPluginTooltips?: boolean;
+
+    channelId?: string;
+
+    /**
+     * Post id prop passed down to markdown image
+     */
+    postType?: PostType;
+    emojiMap?: EmojiMap;
+
+    /**
+     * Some components processed by messageHtmlToComponent e.g. AtSumOfMembersMention require to have a list of userIds
+     */
+    userIds?: string[];
+
+    /**
+     * Some additional data to pass down to rendered component to aid in rendering decisions
+     */
+    messageMetadata?: Record<string, string>;
+}
 
 function Markdown({
     options = {},
@@ -19,22 +90,22 @@ function Markdown({
     postId = '', // Needed to avoid proptypes console errors for cases like channel header, which doesn't have a proper value
     editedAt = 0,
     message = '',
+    channelNamesMap,
+    mentionKeys,
+    imageProps,
+    channelId,
+    hasPluginTooltips,
+    postType,
+    emojiMap,
+    userIds,
+    messageMetadata,
     enableFormatting,
     autolinkedUrlSchemes,
     siteURL,
-    mentionKeys,
-    channelNamesMap,
     hasImageProxy,
     team,
     minimumHashtagLength,
     managedResourcePaths,
-    emojiMap,
-    imageProps,
-    hasPluginTooltips,
-    userIds,
-    messageMetadata,
-    channelId,
-    postType,
 }: Props) {
     if (message === '' || !enableFormatting) {
         return (
